LayoutTests/editing/deleting/ios/backspace-last-character-expected.txt

This tests using the software keyboard and pressing 'w' then backspace deletes the 'w' and does not cause an assertion failure in a debug build.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById("input").value is "w"
PASS document.getElementById("input").value is ""
PASS successfullyParsed is true

TEST COMPLETE
